+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @emberkit/cli
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @emberkit/core@0.9.0
+
 ## 1.0.0
 
 ### Patch Changes